Add files via upload
In this lecture, hosted by Rainer Stropek from the Pädagogische Hochschule Oberösterreich (Linz, Austria), I tried out the basics about HTML and CSS. In order to learn the basic coding, I created a webpage about War Comics. 

The content is based on my wikipedia article Krieg im Comic. Originally, this article was written in summer 2021.
I do not own the content of the text or images of this webpage or wikipedia article and all rights belong the the responsible artists, teachers, researchers and/or their companies. If something is not citied correctly or the provided information is wrong, please contact me. I will make the changes